requirements.txt
asgiref==3.8.1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.12.2
certifi==2024.6.2
charset-normalizer==3.3.2
crispy-bootstrap5==0.6
Django==4.2.13
django-appconf==1.0.5
django-crispy-forms==1.14.0
django-pagedown==2.2.1
django-ranged-response==0.2.0
django-redis==5.2.0
django-rosetta==0.9.9
django-select2==8.1.1
django-simple-captcha==0.6.0
et-xmlfile==1.1.0
graphviz==0.20.1
gunicorn==20.1.0
idna==3.7
iniconfig==1.1.1
lxml==4.8.0
Markdown==3.3.4
openpyxl==3.0.10
packaging==21.3
pillow==10.3.0
pluggy==1.0.0
polib==1.2.0
psycopg2-binary==2.9.4
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
redis==4.5.2
requests==2.28.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.12.2
urllib3==1.26.18
xmltodict==0.13.0
#========
graphene==3.3
graphene-django==3.1.5
djangorestframework==3.14.0
pyjwt==2.8.0
